backend/requirements.txt
# FastAPI core
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart>=0.0.9

# Configuration
pydantic>=2.7.1
pydantic-settings>=2.3.1

# HTTP client
httpx==0.27.0

# JWT validation
PyJWT>=2.8.0
cryptography>=41.0.7

# Supabase Python client
supabase==2.5.0

# SQLAlchemy + asyncpg (install with --only-binary=:all: on Windows)
sqlalchemy>=2.0.30
asyncpg>=0.29.0
alembic==1.13.1

# Scraping (Lightweight httpx approach optimized for Render Free Tier)
beautifulsoup4==4.12.3
lxml>=5.2.2

# Google AI — Gemini (new unified SDK)
google-genai>=1.0.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7
